split db types
